move package_manager_controller to cli-core
