+++ packages/cli-core/package.json
@@ -18,6 +18,7 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "@inquirer/prompts": "^3.0.0"
+    "@inquirer/prompts": "^3.0.0",
+    "execa": "^8.0.1"
   }
 }
