+++ packages/cli-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './prompter/amplify_prompts.js';
 export { COLOR } from './colors.js';
 export * from './printer/printer.js';
+export * from './package-manager-controller/index.js';
