+++ packages/cli-core/src/package-manager-controller/execute_with_logger.test.ts
