+++ packages/cli-core/src/package-manager-controller/index.ts
@@ -1,0 +1,2 @@
+export * from './package_manager_controller.js';
+export * from './package_manager_controller_factory.js';
