+++ packages/cli-core/src/package-manager-controller/npm_package_manager_controller.test.ts
