+++ packages/cli-core/src/package-manager-controller/npm_package_manager_controller.ts
@@ -1,14 +1,14 @@
-import _fsp from 'fs/promises';
-import * as _path from 'path';
 import { existsSync as _existsSync } from 'fs';
+import _fsp from 'fs/promises';
 import { execa as _execa } from 'execa';
-import { executeWithDebugLogger as _executeWithDebugLogger } from '../execute_with_logger.js';
+import * as _path from 'path';
+import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
 import { PackageManagerController } from './package_manager_controller.js';
 
 /**
- * PnpmPackageManagerController is an abstraction around pnpm commands that are needed to initialize a project and install dependencies
+ * NpmPackageManagerController is an abstraction around npm commands that are needed to initialize a project and install dependencies
  */
-export class PnpmPackageManagerController extends PackageManagerController {
+export class NpmPackageManagerController extends PackageManagerController {
   /**
    * constructor
    */
@@ -22,9 +22,9 @@ export class PnpmPackageManagerController extends PackageManagerController {
   ) {
     super(
       projectRoot,
-      'pnpm',
-      'pnpm',
-      ['init'],
+      'npm',
+      'npx',
+      ['init', '--yes'],
       'install',
       fsp,
       path,
