+++ packages/cli-core/src/package-manager-controller/package_manager_controller.ts
@@ -2,9 +2,9 @@ import { existsSync as _existsSync } from 'fs';
 import _fsp from 'fs/promises';
 import { execa as _execa } from 'execa';
 import * as _path from 'path';
-import { LogLevel } from '@aws-amplify/cli-core';
+import { LogLevel } from '../printer/printer.js';
 import { printer } from '../printer.js';
-import { executeWithDebugLogger as _executeWithDebugLogger } from '../execute_with_logger.js';
+import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
 import { DependencyType } from './package_manager_controller_factory.js';
 
 /**
