+++ packages/cli-core/src/package-manager-controller/pnpm_package_manager_controller.test.ts
