+++ packages/cli-core/src/package-manager-controller/pnpm_package_manager_controller.ts
@@ -1,14 +1,14 @@
-import { existsSync as _existsSync } from 'fs';
 import _fsp from 'fs/promises';
-import { execa as _execa } from 'execa';
 import * as _path from 'path';
-import { executeWithDebugLogger as _executeWithDebugLogger } from '../execute_with_logger.js';
+import { existsSync as _existsSync } from 'fs';
+import { execa as _execa } from 'execa';
+import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
 import { PackageManagerController } from './package_manager_controller.js';
 
 /**
- * NpmPackageManagerController is an abstraction around npm commands that are needed to initialize a project and install dependencies
+ * PnpmPackageManagerController is an abstraction around pnpm commands that are needed to initialize a project and install dependencies
  */
-export class NpmPackageManagerController extends PackageManagerController {
+export class PnpmPackageManagerController extends PackageManagerController {
   /**
    * constructor
    */
@@ -22,9 +22,9 @@ export class NpmPackageManagerController extends PackageManagerController {
   ) {
     super(
       projectRoot,
-      'npm',
-      'npx',
-      ['init', '--yes'],
+      'pnpm',
+      'pnpm',
+      ['init'],
       'install',
       fsp,
       path,
