+++ packages/cli-core/src/package-manager-controller/yarn_modern_package_manager_controller.ts
@@ -4,7 +4,7 @@ import { execa as _execa } from 'execa';
 import * as _path from 'path';
 import { LogLevel } from '@aws-amplify/cli-core';
 import { printer } from '../printer.js';
-import { executeWithDebugLogger as _executeWithDebugLogger } from '../execute_with_logger.js';
+import { executeWithDebugLogger as _executeWithDebugLogger } from './execute_with_logger.js';
 import { PackageManagerController } from './package_manager_controller.js';
 import { DependencyType } from './package_manager_controller_factory.js';
