+++ packages/create-amplify/src/create_amplify.ts
@@ -7,13 +7,15 @@
   If customers have a cached version of the create-amplify package, they might execute that cached version even after we publish features and fixes to the package on npm.
  */
 
+import {
+  LogLevel,
+  PackageManagerControllerFactory,
+} from '@aws-amplify/cli-core';
 import { ProjectRootValidator } from './project_root_validator.js';
 import { AmplifyProjectCreator } from './amplify_project_creator.js';
-import { PackageManagerControllerFactory } from './package-manager-controller/package_manager_controller_factory.js';
 import { getProjectRoot } from './get_project_root.js';
 import { GitIgnoreInitializer } from './gitignore_initializer.js';
 import { InitialProjectFileGenerator } from './initial_project_file_generator.js';
-import { LogLevel } from '@aws-amplify/cli-core';
 import { printer } from './printer.js';
 
 const projectRoot = await getProjectRoot();
